clean up options and fix test code
