+++ scripts/test-error-handler.js
@@ -68,7 +68,7 @@ function runTest(name: string, code: string): boolean {
     let options = {
       internalDebug: false,
       mathRandomSeed: "0",
-      onError: errorHandler.bind(null, recover ? "Recover" : "Fail", errors),
+      errorHandler: errorHandler.bind(null, recover ? "Recover" : "Fail", errors),
       serialize: true,
       speculate: true,
     };
