+++ scripts/test-internal.js
@@ -68,7 +68,7 @@ function runTest(name: string, code: string): boolean {
       compatibility: "jsc-600-1-4-17",
       delayUnsupportedRequires: true,
       mathRandomSeed: "0",
-      onError: errorHandler,
+      errorHandler,
       serialize: true,
       speculate: !modelCode,
       sourceMaps: !!sourceMap,
