+++ src/prepack-cli.js
@@ -132,7 +132,7 @@ function run(
       compatibility,
       mathRandomSeed,
       inputSourceMapFilename: inputSourceMap,
-      onError: errorHandler,
+      errorHandler: errorHandler,
       sourceMaps: !!outputSourceMap,
     },
     flags
