+++ src/prepack-node-environment.js
@@ -18,12 +18,12 @@ import { Completion } from "./completions.js";
 import { Value } from "./values";
 import construct_realm from "./construct_realm.js";
 import initializeGlobals from "./globals.js";
-import { getRealmOptions, getSerializerOptions } from "./options";
+import { getRealmOptions, getSerializerOptions } from "./prepack-options";
 import { FatalError } from "./errors.js";
 import initializeBootstrap from "./intrinsics/node/bootstrap.js";
 import initializeProcess from "./intrinsics/node/process.js";
 
-import type { Options } from "./options";
+import type { Options } from "./prepack-options";
 import { defaultOptions } from "./options";
 import type { SourceMap } from "./types.js";
